Review feedback - unify conftest.py files
